refactor(milo): extract shared TabStrip component (EntityTabNav + Resources reuse it)
