+++ app/features/milo/index.ts
@@ -12,14 +12,8 @@ export { ListPage } from './components/page/list-page';
 export { PlainPage } from './components/page/plain-page';
 export { DetailPage } from './components/page/detail-page';
 export { EntityHeader } from './components/page/entity-header';
-export {
-  EntityTabNav,
-  tabStripClass,
-  tabBase,
-  tabActive,
-  tabIdle,
-  type EntityTab,
-} from './components/page/entity-tab-nav';
+export { EntityTabNav, type EntityTab } from './components/page/entity-tab-nav';
+export { TabStrip } from './components/page/tab-strip';
 export { DetailShell } from './components/page/detail-shell';
 export {
   FilterPanel,
